requirements.txt
Django==4.2.20
djangorestframework==3.15.2
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.31.0 